feat: force the use of a "test" NODE_ENV when running jest
Due to recent changes in liferay-portal, we started setting an explicit
NODE_ENV of "production" by default, which meant that Jest's usual
behavior of assuming "test" no longer worked, and tests broke.

We resolved that in:

    https://github.com/liferay/liferay-portal/commit/0741aabb5d201578a0

But in this commit we want to make ourselves 100% resilient to any
future changes in the way `liferay-npm-scripts test` is invoked by
always setting a NODE_ENV of "test", no matter what.

If invoked with no pre-existing NODE_ENV we print the following:

    Using NODE_ENV: "test"

If invoked with a pre-existing NODE_ENV other than "test", we print:

    Using NODE_ENV: "test"

    Jest requires a NODE_ENV of "test", so the pre-existing
    NODE_ENV of "development" was overridden

If invoked with a pre-existing NODE_ENV of "test", we don't print
anything at all.

Test plan: `yarn add` this package in portal and run tests with various
NODE_ENV settings; see them pass.