fix: update/remove bad links and add missing links in package.json files
Mostly so that the links back to the repo from the NPM pages will work:

- https://www.npmjs.com/package/liferay-jest-junit-reporter
- https://www.npmjs.com/package/liferay-npm-bundler-preset-liferay-dev
- https://www.npmjs.com/package/liferay-npm-scripts

I don't think anybody is using the `npm bugs` command on these packages,
so the "bugs" link is pretty useless. In any case, `npm bugs` is smart
enough to tack "/issues" onto the end of the repo URL anyway.