test(reporter): use a more realistic fake path in tests
This makes the test a little more powerful by making sure that the path
that gets process by `formatDirectoryPath()` actually resembles the kind
of path that we'll see when the tests run inside liferay-portal.

We have logic in there that trims off everything up to the end of
"/modules/apps/", so it's useful for our fake path to actually include
"/modules/apps/" in it.