feat: exit with non-zero status when aborting publish
Yesterday I saw a publish fail with this message:

    Failed to automatically publish package due to:

    run(): command `yarn publish --non-interactive --otp 501636` exited with status 1

(I think because my OTP token expired by the time the command actually
got to talking to the registry servers).

However, the exit status of the process itself it still 0, so it is easy
to overlook this message.

If we have an error, let's exit with a status of 1 instead.