fix: search up as high as liferay-portal "root" for user config
We just noticed that when you run the "format" task from a project
directory, it doesn't consult the ".prettierrc.js" at the top level.

This is because we tell cosmiconfig not to walk up above the current
directory. It seems reasonable that we should be able to safely walk up
as high as the "root" (ie. the directory containing the "yarn.lock"
file) though.