Fixes #303: Redesign rules once more
We have been forced to run rules over source files to generate output
files as opposed to directly transform output files as the bundler does
for the rest of artifacts.

This is because transforming in place we don't have enough info to know
which artifacts are clean and which are dirty so we may enter loops
where, after running more than one build without a previous clean, files
are processed more than once and thus get corrupted.

The master plan is to slowly move everything to rules so as to shift the
bundler's running paradigm from "transform in place" to "process sources
to generate outputs". This wasn't possible before because the bundler
delegated several build steps in external build tools, but now we want
the bundler to be in full charge of the build, as webpack does, for
example.

That's the only way to make sure we can optimize to the maximum.