feat!: activate "no-console" rule
Tested in liferay-portal: currently catches 1 issue, so I don't think
this will be very noisy.

Note that I tuned this to not complain about `console.warn` or
`console.error`, which are generally used to report legitimate
problems and are not expected to produce routine console spew (unlike
`console.log`, which is often an overlooked piece of debug code that has
been removed).

Closes: https://github.com/liferay/eslint-config-liferay/issues/74