feat: add support for EL syntax
The old implementation had partial support for this, even though there
are almost no uses at all in liferay-portal. But seeing the spec is
pretty limited (eg. no nesting), and seeing we can use negative
lookbehind in the version of node that we use (v10), it is pretty easy
to implement full support here (ie. including support for escaping).