feat: use shorter expression IDs
Based on count rather than location. Because I want to make these match
the length of the original source, and that will be easier to do for all
cases if the symbols are short.

Do we really need the ID here at all? The original intent was for us
to guarantee uniqueness (see commit below), but I don't know whether
we'd run into that case any more (because ES6 explicitly allows
duplicate keys in objects in order to support spread and override
semantics). Nevertheless, easy enough to keep this around so we may as
well.

    https://github.com/liferay/liferay-frontend-source-formatter/commit/c0d0f8f4a9da46010e96538a1c8a0f832fdefc53