Fixes #393: Let wrap-module-amd plugin use an external template
If babel-plugin-wrap-modules-amd finds a `.wrap-modules-amd.template`
next to the `.js`file it is processing, it will use that as a
`babel-template` to wrap the wrapped module.

This can be used to inject additional JS code that needs to be executed
before the AMD define() happens. Usually to control when that happens.