fix: make offset measurements fully accurate
I don't think this would ever be a problem in practice but may as well
fix it.

Consider replacing these expressions:

    <%= 1 %><%= 2 %>         ... substitute expression at offset 0:
    ʾJSP_EXPRʿ<%= 2 %>       ... string is now 2 longer; then:
    ʾJSP_EXPRʿʾJSP_EXPRʿ     ... second expression now substituted

The "offset" reported for the second position will be the offset in the
original string, not its new position, which is +2 further along. So, if
we want the offset to be accurate, we have to adjust as we go along.