feat: add beginnings of a proper lexer
Because I think the approach taken so far and copied from CSF (repeated
regex replacements over the source) isn't likely to be robust. I think
we're better off doing a real lexer that scans from left to right in one
pass. That way we avoid headaches with things like offsets that move as
we mutate the file and so forth.

Still figuring out what the ideal API should look like for this, so
showing a couple of options so far.

You can either do one `consume()` call and compose matchers inside that;
eg.

    text += consume(sequence(a, anyOf(b, c), repeat(d))).once();

Or you can make multiple calls to `consume()`; eg:

    text += consume(a).once();
    text += consume(anyOf(b, c).once());
    while (peek(d)) {
      text += consume(d).once();
    }

First feels a little more declarative, the second, more imperative. In
practice they are doing exactly the same work. I think I like the idea
of starting declarative and having an imperative layer to drop down to
when necessary.