refactor: simplify `consume()` API
Now that I have written a few productions with the combinator-based API,
I like it better than the `consume()` API, so let's standardize on
that.

This in turn led me to hit a different error path than before, which
obliged me to make a nice `description` for the `until()` combinator,
leading to (I think) more informative error messages. ie:

    Failed to match JSP_COMMENT_START -> JSP_COMMENT_END at: "<%-- unterminated"

instead of:

    Unexpected end-of-input trying to match JSP_COMMENT_END at: ""