feat: lex template text
Note that we have to "peek" farther ahead now to make sure we pick the
right production. So I added memoization for that to make it cheap. The
spec says that all productions are greedy, and if you look at the
"CustomAction" production, you'll see it includes "<", "CustomAction",
and "CustomActionBody" (which itself goes all the way to
"CustomActionEnd", which actually goes all the way to the "ETag"). I
don't think we want to scan that far ahead, if we can help it, so
starting with a relatively minimal lookahead that goes up to the closing
">" or "/>" for now. Seems about right given that we are trying to be a
lexer (tokenize), not a full parser and validator.