test: fix bad fixture
The old `csf` tool used to happily accept JSP tags with unescaped quotes
inside expressions inside attributes, but that isn't valid according to
the spec, and that in turn was causing the new lexer -- which tries hard
to follow the spec as close as is practically possible -- to reject the
fixer.

Fix is to change the fixture, but I added some tests as well to make all
this explicit.