fix: don't choke on stale .babelrc files (#242)
As explained in the related issue, we have a few places around where we
use finally to do cleanup of temporary files we leave on the filesystem.

That deals nicely with errors but if you interrupt a build with Ctrl-C
the finally block(s) won't run. The stale config files (eg. a merged
.babelrc) that get left behind can cause the next run to fail with a
confusing stack trace, usually deep in the bowels of the deepMerge()
function along the lines of target.slice() is not a function.

We should deal with this in depth:

1. Make deepMerge() resilient to being run on a previously merged config.
2. Detect a stale merged config file on disk and report the problem to
   the user in a way that they can understand the fix.
3. Install a sig handler that does clean-up even when a build is
   interrupted.

This PR does 1, and in a way 2 (because if it has a problem merging it
throws an error advising people to check for stale .babelrc file(s)).

I'll follow up with another PR for item 3.

Closes: https://github.com/liferay/liferay-npm-tools/issues/242