fix: don't let signals leave stale files lying around (#242)
Make sure that signals in child process get caught and turned into
errors, so that our `finally` blocks will run and perform their
necessary cleanup even when the user interrupts a build with Ctrl-C.

Originally sent this as:

https://github.com/liferay/liferay-npm-tools/pull/244

But that had two parts, so I am splitting it into:

https://github.com/liferay/liferay-npm-tools/pull/245

And this one so that we can decide independently whether we want both or
just one. This one should clearly go in; the other one may not be
needed but might be nice to have "just in case".

Companion to: https://github.com/liferay/liferay-npm-tools/pull/243