feat: allow for compound patterns in extensions in globs
Back in ancient history before we knew that we were going to have to
implement our own glob expansion, we made a basic `filterGlobs()`
function that would take some globs like:

    ['**/*.js', '**/*.css', '**/*.foo']

And filter them down to a subset:

    ['**/*.js', '**/*.css']

To keep things simple, we didn't add support for globs that used
compound patterns in their extensions (eg. `'**/*.{js,jsx}'`).

Later on, we were forced to implement our own globbing, and part of that
was building a function (`preprocessGlob()`) that could preprocess
complex globs (eg. `'{src,test}/**/*.js'`) and transform them into a
list of simpler globs (eg. `['src/**/*.js', 'test/**/*.js']`).

Looking at the code now, I see that there's no reason to require globs
to have simple extensions any more; we can support compound patterns in
globs by swapping the order of operations from:

- `filterGlobs()` then `preprocessGlob()`; to:
- `preprocessGlob()` then `filterGlobs()`

I made that change in both `format.js` and `lint.js`, and then decided
it was time to refactor them to extract out the common functionality
into a new `getPaths()` module. The only difference between the two
files in the previous implementation was that `format.js` ignored
"node_modules" by default even if not explicitly configured to do so (in
order to match the Prettier behavior). So, the new `getPaths()` function
does this too, seeing as it is harmless to apply the same default
exclusion to ESLint as well.