feat: tweak merge error message to suggest manual review
As per this thread:

https://github.com/liferay/liferay-npm-tools/pull/243#discussion_r324080124

The other fix that went in (#244) means that stale files are pretty
unlikely, and we'll probably only hit this error case if a developer
has manually made a broken ".babelrc" file.

So, drop the word "stale" from the message and suggest manual review
instead. It's hard to provide any more concrete suggestions than that
because this is a recursive function: we could print the unmergeable
object(s) to the console, for example, but that might actually cause
more confusion because it may not be obvious what the objects refer to.