Fixes #407: Normalize all path usages
We leverage FilePath class to achieve this goal. In addition we
normalize variable names following these conventions:

1. FilePath objects are named <something>Dir or <something>File
2. String paths are named <something>Path

For example: sourceDir, sourcePath.

Because paths can be absolute or relative to something (the project
directory, a package directory, the build directory, ...) we will use
prefixes for the Dir, File, and Path suffixes:

1. Any absolute path/FilePath will be prefixed with abs, so for example:
jsonAbsPath or jsonAbsFile.
2. Any relative path/FilePath will be prefixed with <location>Rel where
<location> can be prj, pkg, build, and so on. For example: jsPrjRelPath
or jsPrjRelFile.
3. If none of the abs or <location>Rel prefixes are used, the path is
supposed to be a standard path, i.e., relative to the work directory if
it is not an absolute path.

In the future we may use TypeScript to enforce the use of FilePaths in
every place to avoid mistakes.