refactor: make dedent() return a tuple
This was the way it was only because I originally thought that I would
only need to return a string, then I discovered I needed the tab count
too and didn't want to rewrite the tests, and because there is an
precedent for this ugliness in the form of the `lastIndex` property on
RegExp objects. But really: there is no excuse for this, so let's just
make it a tuple.