refactor: extract "trim()" function
Seeing as all the other steps in `formatJSP()` just call out to
independent functions, we may as well do the same here.

Bonus: we get to write tests for the functionality, and in doing that I
can see that we could actually afford to strip out the matching
whitespace because Prettier will kill it anyway, but I'll do that in a
separate commit to make it clear what's happening.