fix: make "trim()" actually trim
In writing the tests in the previous commit I realized that our "trim"
function wasn't really living up to its name: it was capturing the
leading and trailing whitespace, but not actually removing it.

My original thinking was, no harm done, because Prettier strips it
anyway, but we may as well make the behavior of the function match its
name, to avoid possible confusion.

But it turns out that on testing this in liferay-portal, it actually
fixes whitespace errors in 4 places as seen here:

https://gist.github.com/wincent/817c043bc1184984bdd5541e2e628271

So, adding a test to make sure this doesn't regress.