perf: avoid repeated slice()-ing of entire input string
Old approach was to iterate backwards over the source string splicing in
new segments. Has to go backwards because all of the indexes to the
right of any substitution become invalidated by any change that changes
the string length.

New approach just does one forwards pass instead. Arguably slightly more
complicated because it requires a `lastIndex` book-keeping variable and
must tag on the final source block in the `return` statement, but it is
allocating few strings along the way, and going forwards over the string
is less subtle.