feat: add "next" properties to tokens
In order to adjust whitespace around scriptlet tokens, I'm adding "next"
links to each token (we already had "previous" links), effectively
turning the token stream into a doubly-linked list and making it very
straightforward to implement lookahead and lookbehind.

The cost to implement this is more than just a linked list, because we
have to make sure that any usage of the iterator is unaffected by
lookahead and lookbehind; additionally, there are no guarantees about
what order iteration and link traversal will happen in, or in what
direction, as shown in the tests. In order to be totally flexible, we
have to do some duplicate book-keeping in the form of a map, but the
cost is just some short-lived garbage on the heap.