fix: respect scriptlet-based indentation
For now just doing the simplest thing (inferring implied indent/dedent
by count of braces in each scriptlet); if that proves to be inadequate
we can do a more sophisticated scan.

We could have changed the lookahead that we were doing on each NEWLINE
(to see whether a CLOSE_TAG_REPLACEMENT following that would affect the
indent) to instead look ahead for CLOSE_TAG_REPLACEMENT or SCRIPTLET
tokens.

Seemed simpler, though, to just drop the lookahead and instead handle it
separately in each of CLOSE_TAG_REPLACEMENT and SCRIPTLET; they are
sufficiently different to make handling them separately easier.

One thing that may seem confusing: we both do `removeIndent()` (to strip
off an excess indent that may have been emitted in the WHITESPACE case
before getting to CLOSE_TAG_REPLACEMENT or SCRIPTLET) and also a
`slice()` call. These really cover two separate concerns: one is the
WHITESPACE case mentioned above, and the other is all other cases, where
we may have a pending `indent` still and need to trim it (an example of
such a case would be when the CLOSE_TAG_REPLACEMENT or SCRIPTLET is the
first thing on the line but we still have a pending indent).

Closes: https://github.com/liferay/liferay-npm-tools/issues/259