feat: extract liferay-js-publish package
As explained in the README.md, this is just packaging up the existing
publish.js script so that it can be used from other projects.

Its only dependency is "cross-spawn": 7.0.0 is out and would be a safe
upgrade for us, but it doesn't offer anything that we need (all it does
is drop support for old NodeJS versions), so let's stick with "^6.0.5",
which is what we're already using in liferay-npm-scripts, so as to avoid
unnecessary duplication.

Speaking of duplication, the other thing you'll note is that I have a
copy-paste of "git.js" and "run.js" in here from "liferay-npm-scripts".
It is only a few lines of trivial code, and avoiding the dependency here
is desirable because keeping dependencies in sync, even our own, is a
headache.