chore: update eslint, eslint-config-liferay, prettier
Big update for eslint-config-liferay (from v3.0.0 to v10.0.0), so there
are 437 problems to fix (351 of which are autofixable).

The ESLint update is big too (v5.15.2 to v6.4.0) but that only adds one
more problem, taking the total to 438.

I'm going to split the autofixable fixes and the manual fixes over the
next two commits so that it is clear what they are.

The Prettier update is low key (v1.16.4 to v1.18.2), and it doesn't
introduce any formatting changes.