fix(imports-first): don't let directives cause silent errors
We had a logic error here which was exposed in the presence of a
directive:

     'use strict';

     var x = require('x');

We would walk up looking at tokens and checking them with
`isRequireStatement()`; on reaching the directive the conditional here
was broken, so we'd end up trying to destructure `init` from
`node.declarations`, which was undefined.

So this fixes the bug but leads to some false positive lints because now
they complain about imports not being before directives. Next commit
will add logic to allow those.