feat: add reminder to publish liferay-npm-scripts
Any time we update the reporter or the preset we have to update the
scripts as well, because the latter depends on the first two.

Just insert a little, low-tech reminder to do so.

That wouldn't have stopped me from making the mistake that a just made
though: we'll need something more for that:

1. Made changes to preset and scripts.
2. Published scripts v10.0.0.
3. Realized error, published preset v2.0.0.
4. Updated scripts dependency on preset.
5. Published scripts v10.0.1.

We'll need a bit more logic and automation to catch/prevent that kind of
mistake.