feat: update the no-react-dom-render error message
When we originally added this rule, we didn't have the
`<react:component />` taglib yet. Now, we do, so we should mention it in
the error message. I put it first in the message, before mentioning
`react` from "frontend-js-react-web", because I get the sense we'd like
to encourage use of the taglib.

Closes: https://github.com/liferay/eslint-config-liferay/issues/81