Make TypeScript types integrate with JavaScript
Migrate liferay-npm-build-tools-common, which is the deepest project in
dependency tree to TS. This way, we can generate .d.ts files for it and
start consuming it from other projects.

Note too that liferay-npm-build-tools-common is more or less the
specification of the bundler's API files and plugin interfacing so, now
that we have TS available, we will continue enhancing the type
definitions as a way to document and specify what can and can't be done
and how the JS Toolkit expects to be configured.

In the future we may need to split the project in two: one project with
internal utility tools (like namespace.ts, globs.ts, and so on) and
another with the real exported objects that make part of the exported
API to be consumed by other parties when developing extensions for the
bundler or configuring it from JavaScript (only if we implement that
feature, of course).