Fixes #422: Move package dir detector
Because the old way of detecting the package directory for a random .js
file does not work because it made the incorrect assumption that no
intermediate package.json will be found inside a package, we need to
rewrite it.

The assumption we made know is that we are inside a JS Toolkit project,
thus the algorithm is not going to work for any .js file, just .js files
inside packages inside JS Toolkit projects.

Because of that, and because this algorithm only makes sense for babel
plugins (at least for now) we move the methods to another class.

We also write tests for any possible scenario:

    -Project's package
        -In source folder
        -In build folder
    -Dependency package
        -In project's node_modules folder (source)
        -In build's node_modules
    -Scoped dependency package
        -In project's node_modules folder (source)
        -In build's node_modules