fix(sort-imports): make sure "metal" sorts before "metal-dom"
We had a bug in here due to some special casing for the pathological
scenario in which you have multiple imports from the same module. In
that case, we need to break ties, so we created a sort key of the form
"module-name:tie-breaker" and compared that.

For something like these two imports...

    import dom from 'metal-dom';
    import core from 'metal';

we'd make two sort keys, "metal-dom:dom" and "metal:core", and
unfortunately, "-" sorts before ":", giving us the wrong result.

So, the fix is to only use the tie-breaker if the names are the same.