chore: don't use arrow functions in Liferay.Loader.define calls
We apply some overrides here and undo the related autofixes made in
the previous commit. Specifically, we want to exclude some files
from being modified by the prefer-arrow-functions rule, because we
want the `Liferay.Loader.define()` calls in there to continue using
the `function` keyword so that they match how they actually exist in
liferay-portal.