feat(scripts): teach Jest to deal with new liferay-npm-bundler loaders
As explained in the inline comments, these transforms allow Jest to work
with components that contain SCSS and CSS imports:

     import './MyComponents.scss';

Given such an import Jest will otherwise try to read the file and treat
it as JS, which obviously won't work. We use the transform here to turn
such "modules" into harmless no-ops.