test(scripts): restructure tests
Get rid of unnecessary prior change to the format.js tests and instead
create a dedicated test file for the Prettier wrapper that will enable
us to test our rules more explicitly and conveniently.

One gotcha to note: Prettier is perfectly happy to format "invalid" JS
snippets like this (invalid because we are using `return` outside of a
function):

    if (test) {
        return 1;
    }

ESLint on the other hand silently skips(!) such files, not even invoking
the custom rule at all. So, we need to be very careful when writing
tests: make sure you pass only valid snippets and also confirm that
ESLint is actually doing something (see the test fail first before
changing it and seeing it green).