feat(scripts): make custom formatting work with ES6 files
This is horrible magical voodoo coding here; as noted in the comment, I
don't know why I need to set `parser` to `undefined` here (overriding
"babel-eslint"). Without this, ESLint chokes on the snippets in the
tests (and I imagine it would choke outside of tests too), silently
failing and returning the original input unmodified.

Note that in an earlier commit you can see we successfully made changes
in the JSP fixtures, and the reason why they all worked even without
this change is only because they don't contain any ES6.