test(scripts): prove that we can handle SCSS in the custom formatter
This is mostly in the name of robustness although it works even without
the changes I am making in this commit: if you pass ESLint some random
non-JS "garbage" (ie. SCSS!) it will just return it unchanged without
errors.

What we do here is make things a little more explicit: if we get a
filepath, we check to make sure it is a whitelisted type and proceed
only if it is.

If no filepath is passed, we use a dummy (JS) path. This is ok because
we always pass a filepath for JS and SCSS files; the only files which
don't get passed are JSP files, so the dummy path is appropriate for
those.