fix(scripts): work with ES6 modules
In testing this in liferay-portal, I found that it only updated 244
files out of 4.7k, because it choked on any file containing an "import"
statement.

I believe that using a sourceType of "module" should be fine even when
formatting JSPs (which would more naturally use "script" type), so let's
just do this as a blanket setting; if that assumption later proves to be
incorrect we can override and toggle between "module" and "script"
accordingly at runtime based on the filename.