feat!: assume React by default in liferay-portal (#107)
Before:

| Preset  | Extends                      |
| ------- | ---------------------------- |
| liferay | eslint:recommended, prettier |
| react   | liferay                      |
| metal   | liferay                      |
| portal  | liferay                      |

Now:

| Preset  | Extends                      |
| ------- | ---------------------------- |
| liferay | eslint:recommended, prettier |
| react   | liferay                      |
| metal   | react                        |
| portal  | react                        |

The motivation here is to make it easier to build React projects in
liferay-portal, because right now we have > 30 projects using React and
they all need an ".eslintrc.js" file to select the preset.

After this change, we will be able to delete those special cases, and
replace them with a hopefully-smaller number of config files that select
the "metal" preset instead.

Closes: https://github.com/liferay/eslint-config-liferay/issues/107