feat!: add react/forbid-foreign-prop-types to react preset
This is in service of:

https://github.com/liferay/liferay-npm-tools/issues/301

We want to activate the react-remove-prop-types Babel transform:

https://www.npmjs.com/package/babel-plugin-transform-react-remove-prop-types

And this can help mitigate the associated risks. As per the docs:

https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/forbid-foreign-prop-types.md

> This rule forbids using another component's prop types unless they
> are explicitly imported/exported. This allows people who want to use
> babel-plugin-transform-react-remove-prop-types to remove propTypes
> from their components in production builds, to do so safely.

However, this is still not entirely risk free. The docs recommend the
use of the "import/named" rule, which we don't currently have in our
dependencies (and we are loathe to increase the footprint):

> In order to ensure that imports are explicitly exported it is
> recommended to use the "named" rule in eslint-plugin-import in
> conjunction with this rule.

https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/named.md

The other possible gotcha noted the react/forbid-foreign-prop-types
documentation is this:

> If you are writing a higher-order component that hoists the wrapped
> component's propTypes, you might want to disable this rule.

To my knowledge, we aren't doing this in liferay-portal, but we may be
elsewhere. For the most part, accessing another component's propTypes
sounds like a smell and a violation of encapsulation. My proposal here
is that we proceed here and deal with any fallout in projects via
suppressions, but hopefully we won't need to do anything drastic.