chore: allow `for`/`of` loops
All the environments where we actually run eslint-config-liferay support
`for`/`of` loops properly (ie. recent NodeJS versions). We still want to
keep the rule on by default, until we're sure it is safe to turn it off
in liferay-portal:

https://github.com/liferay/liferay-frontend-guidelines/issues/53#issuecomment-509664883

But it is fine for us to turn it off inside eslint-config-liferay
itself. This gets rid of this error:

    plugins/eslint-plugin-liferay-portal/lib/rules/no-explicit-extend.js
      29:3  error  for..of loops are not allowed  no-for-of-loops/no-for-of-loops