feat(scripts)!: make Babel use preset-react by default (#303)
Note that this will require some corresponding changes in
liferay-portal, namely, there are two projects which use the
incremental-dom plugin:

- https://github.com/liferay/liferay-portal/blob/1f84f9713487704f173e6777781a20d7c38477f6/modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-builder/.babelrc
- https://github.com/liferay/liferay-portal/blob/1f84f9713487704f173e6777781a20d7c38477f6/modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-web/.babelrc

I don't think we can turn off the preset for those, but we can disable
the individual plugins that the preset activates.

I believe all the rest should be fine, but will obviously have to test
things.

Closes: https://github.com/liferay/liferay-npm-tools/issues/303