feat(scripts): provide a way to opt out of default presets (#303)
Babel doesn't provide a way to override a preset (for example, excluding
plugins from it) or to turn it off; their position, since 2015, has been
that if you don't like a preset, you should fork it and create your own:

https://github.com/babel/babel/issues/3016

That approach wouldn't really help us for our use case (which
is to make @babel/preset-react the default, except for two
projects where we use incremental-dom), because even if we made
"babel-preset-liferay-without-react", there would be no way for us to
use it to override the default.

So, we do this ghastly (and hopefully temporary) hack. In the two
projects which need it, we add this their .babelrc:

    {
      "liferay": {
        "excludes": {
          "presets": ["@babel/preset-react"]
        }
      }
    }

In `getMergedConfig()`, we use this info as a cue to strip out
"@babel/preset-react" from any "presets" array in the config. This is
flexible enough to prune out unwanted "presets" or "plugins" anywhere in
the config object (including, for example, nested deeper down under an
"overrides" property). Having said that, we currently only ever set
"presets" at the top level, so the flexibility isn't really needed yet.

Note that we have to pass `--no-babelrc` when we fork the `babel`
process, otherwise it will still try to read the project .babelrc file
and end up throwing an error when it sees the unknown "liferay"
property. We strip out the "liferay" property from the merged config.

Related config changes needed in liferay-portal can be seen here:

https://gist.github.com/wincent/5a8b85809936ce82cf8101f4e75d20af

Test plan: `portool deploy` in a module that uses incremental-dom (eg.
dynamic-data-mapping-form-builder) and one that uses React (eg.
frontend-js-components-web); run the `storybook` script in a project
with a Storybook (eg. portal-search-tuning-rankings-web); run Jest tests
in all of liferay-portal; do an `ant all`.