feat(scripts)!: enforce use of canonical config file names
We want developers to only use `.babelrc.js`, `.eslintrc.js` (and
technically, `.prettierrc.js` too, although there should be exactly one
of those in liferay-portal) so that we can format and lint those files,
to make sure we get the full benefit of rules like this one:

https://github.com/liferay/eslint-config-liferay/blob/master/plugins/eslint-plugin-liferay-portal/docs/rules/no-explicit-extend.md

To that effect, this commit adds an additional "preflight" step to our
"check" script, which means that we now have 3 phases:

1. Preflight checks (currently, these just check for disallowed config
   file names, but this would be a logical spot to put additional checks
   in the future).
2. Formatting checks (ie. Prettier).
3. Lint checks (ie. ESLint).

I initially considered baking the check for config files into the other
steps (ie. the Prettier step could check for bad Prettier config files,
and the lint step could check for bad ESLint config files), but that
would have involved scanning the repo multiple times and I figure it is
better to do one scan (which takes a few seconds) up front. It also ends
up being cleaner because we don't have to juggle multiple error
conditions and produce clear output for them; eg:

- Config is bad but tool succeeded.
- Config is good but tool failed.
- Config is bad and tool failed.
- Config is good and tool succeeded.

So, we just run three steps (preflight, format, lint) and the logic
ends up being very simple.

In any case, we don't have much choice anyway. If we didn't do this,
where would we put the enforcement of the use of ".babelrc.js" files? We
would need to add a separate step for that anyway. As a bonus, this
gives us an obvious extension point for any other non-ESLint checks that
we might want to run in the future.

Test plan: Running this in liferay-portal gives output like this in the
presence of errors:

    $ liferay-npm-scripts check
    Preflight check failed:
    src/.eslintrc: BAD - use .eslintrc.js instead
    src/.prettierrc: BAD - use .prettierrc.js instead

    Prettier checked 11 files

    1 of 3 jobs failed
    error Command failed with exit code 1.

And in a clean run:

    $ liferay-npm-scripts check
    Prettier checked 11 files

    ✨  Done in 1.70s.

Closes: https://github.com/liferay/liferay-npm-tools/issues/339