fix(scripts): teach jest wrapper to find .babelrc.js files
Noticed this in final testing of:

https://github.com/brianchandotcom/liferay-portal/pull/83438

Test runs were failing in two modules (the only ones with .babelrc.js
files) with this message:

    Multiple configuration files found. Please remove one:
     - .babelrc
     - .babelrc.js

The problem is this code in the Jest wrapper that was writing a
duplicate .babelrc file into the correspond directories. In this commit
we teach it to use ".babelrc.js" instead.