refactor(scripts): eliminate many yarn peer dependency warnings (#336)
This commit doesn't address the entire problem described in #336 but it
is an initial start: namely, it reduces the number of peer dependency
warnings that you see when doing a Yarn operation in this repo itself
from 93 to 22 warnings.

As you can see in the lockfile, we're not actually changing the version
of `@babel/core` that we use here (7.8.3); we're just making explicit
the one that we were already using as a result of our transitive
dependencies.

In liferay-portal itself, we end up using 7.5.5, so when we update
liferay-portal, we'll end up updating that to 7.8.3, which I believe
should be a safe change:

    => Found "@babel/core@7.5.5"
    info Reasons this module exists
      - "workspace-aggregator-7751601e-168c-4244-9341-f549a227cad6" depends on it
      - Hoisted from "_project_#liferay-npm-scripts#babel-jest#@jest#transform#@babel#core"
      - Hoisted from "_project_#liferay-npm-scripts#@storybook#react#@svgr#webpack#@babel#core"
      - Hoisted from "_project_#liferay-npm-scripts#@storybook#react#babel-plugin-react-docgen#react-docgen#@babel#core"
      - Hoisted from "_project_#liferay-npm-scripts#jest#jest-cli#jest-config#@babel#core"
      - Hoisted from "_project_#liferay-npm-scripts#@storybook#react#@svgr#webpack#@svgr#plugin-jsx#@babel#core"

Before (93 warnings):

https://gist.github.com/wincent/18b13750a3db2e2c5d39e8e0a214ed28

After (22 warnings):

https://gist.github.com/wincent/f60645483c0ad0d49549d47796996750

Test plan:

1.  Perform update in liferay-portal:

    yarn add -W --dev ~/code/liferay-npm-tools/packages/liferay-npm-scripts

2.  Verify that duplicate versions go away with a little bit of help:

    npx yarn-deduplicate yarn.lock

Related: https://github.com/liferay/liferay-npm-tools/issues/336