refactor(scripts): eliminate remaining peer dependency warnings (#336)
This commit eliminates the last of the peer dependency warnings that get
produced when doing Yarn operations in this repo. As before, note that
nothing substantive changed in the lockfile.

Note that I couldn't do loose versions here because that would have
triggered a bunch of other changes that I didn't want to occur. I think
all of these will go into liferay-portal without causing much
disruption: we currently have unique versions of `@storybook/core` and
`@testing-library/dom`; `regenerator-runtime` is a mess though, without
about 6 versions in the tree there (and even here, there are two, with
an older one being dragged in by `babel-runtime`). In any case, I chose
the latest version of `regenerator-runtime`, 0.13.3 because that is the
one that is currently "winning" in this repo and getting hoisted to the
top (in liferay-portal, 0.12.1 is "winning" at the moment, but upgrading
it should be pretty harmless in any case).

Before (3 warnings):

https://gist.github.com/wincent/c7ec146279dedee6aa9ac349406c7da7

After (0 warnings):

`/dev/null`