refactor: turn off no-for-of-loops in this repo (#344)
In service of #344 I will need to start using async/await in
liferay-npm-scripts (because stylelint only offer a Promise-based API,
and it is going to get real messy real fast without async/await).

In the places where liferay-npm-scripts runs we already have a version
of Node.js that supports this:

- In Liferay DXP 7.2+, Node.js v10.
- In Liferay DXP 7.1, Node.js v8.

We are unlikely to port these liferay-npm-scripts changes back to
Liferay DXP 7.0 -- and if we do, it would be easy enough to add a build
step, or perhaps upgrade the Node.js version there, which is something
we'd like to see anyway -- so this looks like a safe change.

If we're going to assume async/await, we may as well assume `for`/`of`
as well and disable this lint rule in this repo.

Related: https://github.com/liferay/liferay-npm-tools/issues/344